+++ src/Futhark/Compiler.hs
@@ -45,7 +45,9 @@ data FutharkConfig = FutharkConfig
     -- | If true, error on any warnings.
     futharkWerror :: Bool,
     -- | If True, ignore @unsafe@.
-    futharkSafe :: Bool
+    futharkSafe :: Bool,
+    -- | Additional functions that should be exposed as entry points.
+    futharkEntryPoints :: [Name]
   }
 
 -- | The default compiler configuration.
@@ -55,7 +57,8 @@ newFutharkConfig =
     { futharkVerbose = (NotVerbose, Nothing),
       futharkWarn = True,
       futharkWerror = False,
-      futharkSafe = False
+      futharkSafe = False,
+      futharkEntryPoints = []
     }
 
 -- | Print a compiler error to stdout.  The 'FutharkConfig' controls
@@ -122,7 +125,9 @@ runPipelineOnProgram config pipeline file = do
   when (pipelineVerbose pipeline_config) $
     logMsg ("Reading and type-checking source program" :: String)
   (prog_imports, namesrc) <-
-    handleWarnings config $ (\(a, b, c) -> (a, (b, c))) <$> readProgram file
+    handleWarnings config $
+      (\(a, b, c) -> (a, (b, c)))
+        <$> readProgram (futharkEntryPoints config) file
 
   putNameSource namesrc
   when (pipelineVerbose pipeline_config) $
@@ -150,14 +155,15 @@ typeCheckInternalProgram prog =
 -- | Read and type-check a Futhark program, including all imports.
 readProgram ::
   (MonadError CompilerError m, MonadIO m) =>
+  [I.Name] ->
   FilePath ->
   m (Warnings, Imports, VNameSource)
-readProgram = readLibrary . pure
+readProgram extra_eps = readLibrary extra_eps . pure
 
 -- | Not verbose, and terminates process on error.
 readProgramOrDie :: MonadIO m => FilePath -> m (Warnings, Imports, VNameSource)
 readProgramOrDie file = liftIO $ do
-  res <- runFutharkM (readProgram file) NotVerbose
+  res <- runFutharkM (readProgram mempty file) NotVerbose
   case res of
     Left err -> do
       dumpError newFutharkConfig err